Shin woon

+++ peep/lib/home/emotion_chart.dart
@@ -29,7 +29,7 @@ class EmotionChart extends StatelessWidget {
 class ChartData {
   ChartData(this.x, this.y, [this.color]);
   final String x;
-  final double y;
+  final int y;
   final Color color;
 
 }
